Exponentiation
Input
The input will consist of a set of pairs of values for R and n. The R
value will occupy columns 1 through 6, and the n value will be in
columns 8 and 9.

Output
The output will consist of one line for each line of input giving the
exact value of R^n. Leading zeros should be suppressed in the output.
Insignificant trailing zeros must not be printed. Don't print the
decimal point if the result is an integer.